requirements.txt
pandas==1.4.1
scikit-learn==1.0.2
pyspellchecker==0.6.3